Seperated header to its own component and changed the design to make it differ more from the next website
