+++ src/styles/Home.css
@@ -20,7 +20,7 @@ body {
   width: 100%;
   height: 40px;
   background-color: #f9fafb;
-  color: #565656;
+  color: #6a6a6a;
   text-align: left;
   padding: 2px 150px;
   font-family: var(--ff-chrome);
@@ -40,7 +40,7 @@ body {
 }
 
 .main-header .subtitle {
-  font-size: 11px;
+  font-size: 10px;
   line-height: 0.1;
 }
 
@@ -48,9 +48,9 @@ body {
 /* Väder */
 
 .main-header .weather {
-  font-size: 11px;
+  font-size: 10px;
   line-height: 0.1;
-  color: #666;
+  color: #8b8787;
 }
 
 .weather {
@@ -58,7 +58,7 @@ body {
 }
 
 .main-header .weather-location {
-  font-size: 11px;
+  font-size: 10px;
   line-height: 0.1;
 }
 
@@ -93,14 +93,6 @@ body {
   background-color: #ffeb18;
   color: #222222;
 }
-
-
-
-
-
-
-
-
 
 /* Två kolumner, datum och väder*/
 .header-left, .header-right, .header-icon, .header-temp {
@@ -138,7 +130,7 @@ body {
 .header-temp {
   color: #222;
   font-family: Verdana, sans-serif;
-  font-size: 16px;
+  font-size: 12px;
   font-weight: 700;
   line-height: 18px;
   padding-left: 0px;
@@ -154,7 +146,7 @@ body {
   text-align: left;
   font-family: var(--ff-chrome);
   padding: 16px;
-  font-size: 13px;
+  font-size: 12px;
   font-weight: 600;
   line-height: 14px;
   z-index: 999;
@@ -274,7 +266,7 @@ body {
 
 .time {
   font-family: Arial, Helvetica, Verdana, Geneva, sans-serif;
-  color: #909090;
+  color: #a8a8a8;
   line-height: 1.2;
   margin-right: 5px;
 }
@@ -292,7 +284,7 @@ body {
 
 .news-text {
   font-family: Verdana, Arial, Helvetica, sans-serif;
-  font-size: 15px;
+  font-size: 14px;
   color: #222222;
   line-height: 1.2;
   letter-spacing: -0.8px;
@@ -339,9 +331,9 @@ body {
 }
 
 .news-description {
-  font-size: 16px;
+  font-size: 12px;
   margin: 4px 0 0 0;
-  color: #444444;
+  color: #666;
   font-family: Arial, Helvetica, Verdana, Geneva, sans-serif;
 }
 
@@ -359,8 +351,8 @@ body {
 }
 
 .big-news-one p {
-  font-size: 14px;
-  color: #666;
+  font-size: 12px;
+  color: #8d8d8d;
 }
 
 .breaking-news-one {
@@ -370,6 +362,11 @@ body {
   font-weight: bolder;
   font-family: Verdana, Arial, Helvetica, sans-serif;
   line-height: 1.08;
+}
+
+.big-news p {
+  font-size: 14px;
+  color: #8d8d8d;
 }
 
 .headline-one {
@@ -384,7 +381,7 @@ body {
 }
 
 .news-description-one {
-  font-size: 16px;
+  font-size: 10px;
   margin: 4px 0 0 0;
   color: #444444;
   font-family: Arial, Helvetica, Verdana, Geneva, sans-serif;
@@ -411,6 +408,10 @@ body {
   flex-wrap: wrap;
   padding: 2px 150px;
   gap: 30px; 
+}
+
+.menu-content li {
+  font-size: 10px;
 }
 
 .menu-section {
