+++ src/styles/Support.css
@@ -13,7 +13,7 @@ body {
     left: 0;
     height: 270px;
     background-color: rgb(221,42,48);
-    color: #FFFFFF;
+    color: #b7b7b7;
     text-align: left;
     padding: 2px 150px;
     font-family: Arial, Helvetica, Verdana, Geneva, sans-serif;
@@ -22,7 +22,7 @@ body {
 
 .contact-title {
     font-size: 34PX;
-    color: #FFFFFF;
+    color: #dedede;
     font-weight: 600;
     font-family: Arial, Helvetica, Verdana, Geneva, sans-serif;
     text-align: center;
@@ -31,7 +31,7 @@ body {
 }
 
 .contact-text {
-    font-size: 16px;
+    font-size: 12px;
     color: #000000;
     font-weight: 400;
     text-align: center;
@@ -42,8 +42,8 @@ body {
 
 .search-container {
     display: flex;
-    justify-content: center; /* Centrerar sökrutan */
-    margin-top: 30px; /* Avstånd från rubriken */
+    justify-content: center;
+    margin-top: 30px;
 }
 
 .search-box {
@@ -55,7 +55,7 @@ body {
     border-radius: 20px;
     outline: none;
     color: rgba(0, 0, 0, 0.8);
-    font-size: 18px;
+    font-size: 14px;
 }
   
 .support-side {
@@ -87,15 +87,15 @@ body {
 }
 
 .support-text {
-    color: #020B23;
-    font-size: 23px;
+    color: #4c4c4c;
+    font-size: 20px;
     font-style: normal;
     font-weight: 700;
     font-family: Arial, Helvetica, Verdana, Geneva, sans-serif;
 }
 
 .support-side p {
-    color: #000000;
+    color: #706f6f;
     font-family: Arial, Helvetica, Verdana, Geneva, sans-serif;
     font-size: 16px;
     line-height: 1.5;
